sxhkd: remove sxhkd service and make it as a daemon
This is better because the daemon can access all environment variables
automatically. The systemd service runs isolated from the shell and for
accessing environment variables such as devices addresses and API keys (e.g.
lights on/off, sonoff scripts, etc), I have to specify explicitly all
the environemt variables or environemt files. And this is not very
practical.